+++ CHANGELOG.md
@@ -6,7 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 <!-- markdownlint-disable no-duplicate-heading -->
-## v0.3.0
+
+## v0.3.1 (2024-07-25)
+
+### Added
+
+- `CircledNumber` is a new system that uses Unicode circled numbers instead of
+  decimal digits. Thank you to @ModProg for this contribution!
+
+## v0.3.0 (2024-01-19)
 
 ### Breaking Changes